+++ controller/accounts.go
@@ -118,6 +118,8 @@ func (a *Accounts) Update(input model.UpdateAccount) (*model.Account, error) {
 		}
 	}
 
+	adaptedCode := contentAddressesFromInput(*input.DeployedCode)
+	input.DeployedCode = &adaptedCode
 	account, err = a.blockchain.DeployContract(input.ProjectID, acc.Address, *input.DeployedCode)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to deploy account code")
